feat(probes): keep raw temperatures at full precision (no int truncation)
The Kalman filter is now applied centrally to each probe's raw output, so
truncating readings to integers before filtering needlessly quantizes the
input. Replace int() truncation of raw temperatures with float-preserving
conversions:

- Fahrenheit conversion in read_all_ports now rounds to 1 decimal (matching
  the Celsius precision) instead of int(): max31856, mcp9600, max31865,
  max31865_adafruit, ds18b20.
- base._to_fahrenheit() rounds to 1 decimal instead of int() (used by the
  iBBQ, Meater, and ThermoWorks Cloud probes).
- Meater native count/ambient conversions and F read path drop int()
  truncation so precision survives to the downstream rounding.

+++ probes/bt_meater.py
@@ -101,7 +101,7 @@ class BaseMeater:
 		tip = self.bytesToInt(array[0], array[1])
 		ra = self.bytesToInt(array[2], array[3])
 		oa = self.bytesToInt(array[4], array[5])
-		return int(tip + (max(0, (((ra - min(48, oa)) * 16) * 589) / 1487)))
+		return tip + (max(0, (((ra - min(48, oa)) * 16) * 589) / 1487))
 
 	def readCharacteristic(self, c):
 		try:
@@ -224,7 +224,7 @@ class MeaterPro(BaseMeater):
 		return struct.unpack_from('<h', data, offset)[0]
 
 	def ambient_correction(self, ambient_temp, internal_temp):
-		return int(internal_temp + ((ambient_temp - internal_temp) * 1.2))
+		return internal_temp + ((ambient_temp - internal_temp) * 1.2)
 
 	def convert_to_temperatures(self, data):
 		self.internal_temps = [
